add mybatis-gererator, Error in previous merge

+++ src/main/java/com/sdut/community/model/Question.java
@@ -13,7 +13,7 @@ public class Question {
     private String tags;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
